# docs/openapi.yaml
...
openapi: 3.0.4
info:
  title: dployr Base API
  version: 1.0.0
  description: |
    Your app, your server, your rules!

    ## Authentication
    Authenticate via OAuth2.0 (Google, GitHub, Microsoft) or email OTP.
    Sessions are cookie-based.

    ## Authorization
    Role-based access control:
    - **Owner**: Full control, manages admins and billing
    - **Admin**: User management, infrastructure
    - **Developer**: Deploy apps, manage integrations
    - **Viewer**: Read-only access

    ## Rate Limiting
    - Global: 100 requests/minute per user or IP on `/v1/*`
    - Strict: 10 requests/minute on sensitive endpoints
    - Returns `429` with `Retry-After` and `X-RateLimit-*` headers on breach
servers:
  - url: https://base.dployr.dev
    description: Development
  - url: https://base.dployr.io
    description: Production
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: Resource not found
            code:
              type: string
              description: |
                Machine-readable error code. Known values:
                - `request.missing_params`
                - `request.bad_request`
                - `request.invalid_otp`
                - `request.unprocessable_entity`
                - `request.too_many_requests`
                - `request.internal_server_error`
                - `validation.missing_fields`
                - `validation.invalid_format`
                - `auth.bad_session`
                - `auth.bad_token`
                - `auth.invalid_oauth_provider`
                - `auth.bad_oauth_state`
                - `permission.forbidden`
                - `permission.owner_role_required`
                - `permission.admin_role_required`
                - `permission.developer_role_required`
                - `permission.viewer_role_required`
                - `resource.missing_resource`
                - `resource.conflict`
                - `runtime.bad_webhook_signature`
                - `runtime.internal_server_error`
                - `runtime.node_not_connected`
                - `bootstrap.bootstrap_setup_failure`
                - `bootstrap.bootstrap_run_failure`
              example: resource.missing_resource
            helpLink:
              type: string
              format: uri
              example: https://status.dployr.io
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          nullable: true
          type: object
...
paths:
...
  /v1/auth/login/{provider}:
    get:
      operationId: initiateOAuthLogin
      tags: [Auth]
      summary: Initiate OAuth login
      security: []
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [google, github, microsoft]
        - in: query
          name: redirect_to
          schema:
            type: string
      responses:
        '302':
          description: Redirect to OAuth provider
        '400':
          description: Invalid provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/auth/callback/{provider}:
    get:
      operationId: oauthCallback
      tags: [Auth]
      summary: OAuth callback
      security: []
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [google, github, microsoft]
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to app with session cookie set
        '400':
          description: Invalid state or missing parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/auth/login/email:
    post:
      operationId: requestEmailOtp
      tags: [Auth]
      summary: Request email OTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          email:
                            type: string
                            format: email
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/auth/login/email/verify:
    post:
      operationId: verifyEmailOtp
      tags: [Auth]
      summary: Verify email OTP and create session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                  minLength: 6
                  maxLength: 6
      responses:
        '200':
          description: Login successful, session cookie set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid code format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  message: Invalid or expired code
                  code: request.invalid_otp
  /v1/auth/logout:
    get:
      operationId: logout
      tags: [Auth]
      summary: Logout and clear session
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /v1/proxy/node/routes:
    post:
      operationId: addProxyNodeRoute
      tags: [Proxy]
      summary: Add a proxy route on a node (fire-and-return)
      description: |
        Dispatches a proxy route add task to the target node and returns immediately.
        Requires developer role on the cluster.
      security:
        - sessionCookie: []
      parameters:
        - in: query
          name: clusterId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instanceName, serviceName, upstream]
              properties:
                instanceName:
                  type: string
                  description: Target instance tag/name
                serviceName:
                  type: string
                upstream:
                  type: string
                  description: Upstream address (host:port)
                domain:
                  type: string
                  nullable: true
                  description: Optional custom domain
                template:
                  type: string
                  nullable: true
                  description: Optional proxy template name
      responses:
        '202':
          description: Route add task dispatched
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          taskId:
                            type: string
                          message:
                            type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions (developer role required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Node not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/proxy/node/routes/{service}:
    delete:
      operationId: deleteProxyNodeRoute
      tags: [Proxy]
      summary: Delete a proxy route on a node (fire-and-return)
      description: |
        Dispatches a proxy route removal task to the target node and returns immediately.
        Requires developer role on the cluster.
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: service
          required: true
          schema:
            type: string
          description: Service name whose route should be removed
        - in: query
          name: instanceName
          required: true
          schema:
            type: string
          description: Target instance tag/name
        - in: query
          name: clusterId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Route removal task dispatched
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          taskId:
                            type: string
                          message:
                            type: string
        '400':
          description: Missing instanceName or service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions (developer role required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Node not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── Status (Wake-on-demand) ────────────────────────────────────────────────
...